cleaned up home.js by separating it into several components
