+++ study-buddy-frontend/src/components/Navbar.module.css
@@ -10,3 +10,8 @@
 .nav h3 {
     margin-right: 20px;
 }
+
+.whiteButton {
+    color: white !important; /* Ensures that the text color is white */
+    border-color: white !important; /* For outlined buttons, ensures the border color is white */
+}
